Sygus grammar reduce utility.
